+++ app.py
@@ -2,7 +2,7 @@ import json
 import requests
 import urllib3
 import sys
-import re
+# import re
 import os
 import time
 # import tzlocal
@@ -24,26 +24,25 @@ apipass = os.environ.get('APIPASS')
 DISCORD_ID = os.environ.get('DISCORD_ID')
 DISCORD_TOKEN = os.environ.get('DISCORD_TOKEN')
 
+# Data sources
+ISN_URL = 'https://services.swpc.noaa.gov/json/f107_cm_flux.json'
+KP_URL = 'https://services.swpc.noaa.gov/json/planetary_k_index_1m.json'
+
+# Splunk test url and headers
+splunkurl = 'https://splunk.home:8088/services/collector/event?sourcetype=isn_daily'
+splunkheaders = {
+    'Authorization': 'Splunk 0dffc34c-268f-4b38-88a2-9f3e67498f2c'}
+
+# Splunk index
+index = 'isn'
+
+# Elasticearch params & headers
+esurl = 'http://elasticsearch.home:9200/solar/_doc/'
+esheaders = {'Content-Type': 'application/json'}
+
 
 app = FastAPI()
 security = HTTPBasic()
-
-# APM
-# from elasticapm.contrib.flask import ElasticAPM
-
-# app.config['ELASTIC_APM'] = {
-#   # Set required service name. Allowed characters:
-#   # a-z, A-Z, 0-9, -, _, and space
-#   'SERVICE_NAME': 'nova2046',
-
-#   # Use if APM Server requires a token
-#   'SECRET_TOKEN': '',
-
-#   # Set custom APM Server URL (default: http://localhost:8200)
-#   'SERVER_URL': 'http://apm.home:8200',
-# }
-# apm = ElasticAPM(app)
-# APM
 
 
 def getAuth(credentials: HTTPBasicCredentials = Depends(security)):
@@ -60,42 +59,11 @@ def getAuth(credentials: HTTPBasicCredentials = Depends(security)):
         )
     return credentials.username
 
-# import blynklib
-# import threading
-
 
 # Blynk threaded setup
 # Blynk auth token
 BLYNK_API_URL = os.environ.get("BLYNK_API_URL")
 BLYNK_TOKEN = os.environ.get("BLYNK_TOKEN")
-
-# Init blynk
-# blynk = blynklib.Blynk(BLYNK_TOKEN, server="192.168.1.100", port=8081)
-
-# APP_CONNECT_PRINT_MSG = '[APP_CONNECT_EVENT]'
-# APP_DISCONNECT_PRINT_MSG = '[APP_DISCONNECT_EVENT]'
-
-# @blynk.handle_event('internal_acon')
-# def app_connect_handler(*args):
-#     print(APP_CONNECT_PRINT_MSG)
-
-
-# @blynk.handle_event('internal_adis')
-# def app_disconnect_handler(*args):
-#     print(APP_DISCONNECT_PRINT_MSG)
-
-# create blynk loop that will run as own thread
-# def blynkLoop():
-# 	global blynk
-# 	print("Blynk thread Started")
-# 	#blynk.set_user_task(my_user_task, 1000)
-# 	blynk.run()
-
-# Start blynk background thread
-# t = threading.Thread(target = blynkLoop, daemon = True)
-# t = threading.Thread(target = blynkLoop)
-# t.start()
-# Blynk threaded setup
 
 # Set up logging
 logging.basicConfig(
@@ -104,15 +72,26 @@ logging.basicConfig(
 # Disable https warning
 urllib3.disable_warnings()
 
+# Extract version from package.json and print to console
+try:
+    with open('package.json', 'r') as f:
+        package_data = json.load(f)
+        version = package_data.get('version', 'Unknown')
+        print(f"Version: {version}")
+except FileNotFoundError:
+    logging.info("package.json not found.")
+except json.JSONDecodeError:
+    logging.info("Failed to decode package.json.")
+
 
 def dailyIsnIngest():
     logging.info('---  ingest scheduler start ---')
-    page = get_page(isn_url)
-    isn, flux_10cm, Kp, epoch, date = extract_data(page)
+    data = get_data(ISN_URL)
+    isn, flux_10cm, Kp, epoch, date = extract_isn(data)
 
-    logging.info('---  Sending to Splunk ---')
-    resp = hec_send(isn, flux_10cm, Kp, epoch)
-    logging.info('Splunk response: %s', resp)
+    # logging.info('---  Sending to Splunk ---')
+    # resp = hec_send(isn, flux_10cm, Kp, epoch)
+    # logging.info('Splunk response: %s', resp)
 
     logging.info('---  Sending to Elastic ---')
     resp = es_send(isn, flux_10cm, Kp, date)
@@ -126,18 +105,22 @@ def dailyIsnIngest():
 
 def dailyIsnDiscord():
     logging.info('--- Discord scheduler start ---')
-    page = get_page(isn_url)
-    isn, flux_10cm, Kp, epoch, date = extract_data(page)
+    ISN_DATA = get_data(ISN_URL)
+    KP_DATA = get_data(KP_URL)
+    isn, flux_10cm, epoch, date = extract_isn(ISN_DATA)
+    kp, epoch, date = extract_kp(KP_DATA)
 
-    resp = discord_post(isn, flux_10cm, Kp)
+    resp = discord_post(isn, flux_10cm, kp)
     logging.info('Discord: %s', resp)
     logging.info('---  Discord scheduler end  ---')
 
 
 def postBlynk():
     logging.info('--- Blynk scheduler start ---')
-    page = get_page(isn_url)
-    isn, flux_10cm, Kp, epoch, date = extract_data(page)
+    ISN_DATA = get_data(ISN_URL)
+    KP_DATA = get_data(KP_URL)
+    isn, flux_10cm, epoch, date = extract_isn(ISN_DATA)
+    kp, epoch, date = extract_kp(KP_DATA)
 
     # blynk.virtual_write(31, isn)
     # blynk.virtual_write(32, flux_10cm)
@@ -145,7 +128,7 @@ def postBlynk():
 
     def pinurl(pin, metric):
         pinUrl = BLYNK_API_URL + '/' + BLYNK_TOKEN + \
-            '/update/' + pin + '?value=' + metric
+            '/update/' + pin + '?value=' + str(metric)  # Convert metric to string
         return pinUrl
 
     pinUrl = pinurl('V31', isn)
@@ -154,93 +137,103 @@ def postBlynk():
     pinUrl = pinurl('V32', flux_10cm)
     requests.get(pinUrl, verify=False)
 
-    pinUrl = pinurl('V33', Kp)
+    pinUrl = pinurl('V33', kp)
     requests.get(pinUrl, verify=False)
 
     logging.info('Posted to Blynk: ISN, Flux, Kp; %s %s %s',
-                 isn, flux_10cm, Kp)
+                 isn, flux_10cm, kp)
     logging.info('---  Blynk scheduler end  ---')
 
 
-sched = BackgroundScheduler(daemon=True)
-# Schedule the ingestion every 6 hours
-sched.add_job(func=dailyIsnIngest, trigger='interval', hours=6)
-# Schedule posting to discord every 12 hours
-sched.add_job(func=dailyIsnDiscord, trigger='interval', hours=24)
-# Schedule posting to Blynk
-sched.add_job(func=postBlynk, trigger='interval', minutes=29)
-# start all schedulers
-sched.start()
-# Log all sched jobs
-logging.info(sched.print_jobs(out=sys.stdout))
-
-# app = Flask(__name__)
-# app = FastAPI()
-
-# Url to get ISN data from
-isn_url = 'http://sunspotwatch.com'
-
-# Splunk test url and headers
-splunkurl = 'https://splunk.home:8088/services/collector/event?sourcetype=isn_daily'
-splunkheaders = {
-    'Authorization': 'Splunk 0dffc34c-268f-4b38-88a2-9f3e67498f2c'}
-
-# Splunk index
-index = 'isn'
-
-# Elasticearch params & headers
-esurl = 'http://elasticsearch.home:9200/solar/_doc/'
-esheaders = {'Content-Type': 'application/json'}
-
-# Get webpage
-
-
-def get_page(url):
-    logging.info('Trying to get metrics from page...')
+def get_data(url):
+    logging.info('Getting data...')
     try:
         r = requests.get(url)
         r.raise_for_status()
-        page = r.text
-        logging.info('Page %s loaded ok. %s', url, r)
-        return page
+        data = r.json()
+        logging.info('Data from %s loaded ok. %s', url, r)
+        return data
     except requests.exceptions.HTTPError as err:
-        logging.error('Page %s load failed. %s Error: %s', url, r, err)
+        logging.error('Data %s load failed. %s Error: %s', url, r, err)
     except Exception as e:
-        logging.error('Page load, some other error: %s', e)
+        logging.error('Data load, some other error: %s', e)
 
 
-def extract_data(page):
+def extract_kp(data):
     try:
-        logging.info('Extracting page data...')
-        match = re.search(
-            r"Sun.Spots\:\s\<b\>(\d+)\<\/b\>.as.of\s(\d+\/\d+\/\d+)", page)
-        isn = match.group(1)
-        date = match.group(2)
-        match = re.search(r"10\.7\-cm\sFlux\:\s\<b\>(\d+)", page)
-        flux_10cm = match.group(1)
+        logging.info('Extracting Kp-index data...')
+        if not data:
+            raise ValueError("No data found in the response.")
 
+        # Sort data by time_tag to get the most recent entry
+        data.sort(key=lambda x: x['time_tag'], reverse=True)
+        latest_entry = data[0]
+
+        kp = latest_entry.get('kp_index', None)
+        if kp is None:
+            raise ValueError("Kp value not found in the response.")
+        # Convert kp to an integer
         try:
-            match = re.search(
-                r"Planetary K-index\s\(\<b\>Kp\<\/b\>\)\:\s\<b\>(\d+)", page)
-            Kp = match.group(1)
-        except Exception as e:
-            logging.error('Could not extract KP, setting to 0, {}'.format(e))
-            Kp = str(0)
-
-        date_object = datetime.strptime(date, '%m/%d/%Y')
+            kp = int(kp)
+        except ValueError as e:
+            logging.error('Failed to convert kp to integer: %s', e)
+            raise
+        time_tag = latest_entry['time_tag']
+        date_object = datetime.strptime(time_tag, '%Y-%m-%dT%H:%M:%S')
         epoch = time.mktime(date_object.timetuple())
         esdate_object = date_object.strftime('%Y-%m-%dT%H:%M:%S')
-        logging.debug('Daily ISN Count: %s', isn)
-        logging.debug('Daily 10.7cm Flux: %s', flux_10cm)
-        logging.debug('Daily Kp Index: %s', Kp)
-    except AttributeError as e:
-        logging.error('Attribute Error: %s', e)
-    except UnboundLocalError as e:
-        logging.error('Unbound Local Error: %s', e)
+
+        logging.info('Kp Index: %s', kp)
+    except json.JSONDecodeError as e:
+        logging.error('JSON decode error: %s', e)
+    except ValueError as e:
+        logging.error('Value Error: %s', e)
     except Exception as e:
         logging.error('Extract data, other error: %s', e)
     else:
-        return isn, flux_10cm, Kp, epoch, esdate_object
+        return kp, epoch, esdate_object
+        # return kp
+
+
+def extract_isn(data):
+    try:
+        logging.info('Extracting 10.7cm Flux data...')
+        if not data:
+            raise ValueError("No data found in the response.")
+
+        # Sort data by time_tag to get the most recent entry
+        data.sort(key=lambda x: x['time_tag'], reverse=True)
+        latest_entry = data[0]
+
+        flux_10cm = latest_entry.get('flux', None)
+        if flux_10cm is None:
+            raise ValueError("Flux value not found in the response.")
+        # Convert flux_10cm to an integer
+        try:
+            flux_10cm = int(flux_10cm)
+        except ValueError as e:
+            logging.error('Failed to convert flux_10cm to integer: %s', e)
+            raise
+        # Extract other necessary values (if needed)
+        time_tag = latest_entry['time_tag']
+        date_object = datetime.strptime(time_tag, '%Y-%m-%dT%H:%M:%S')
+        epoch = time.mktime(date_object.timetuple())
+        esdate_object = date_object.strftime('%Y-%m-%dT%H:%M:%S')
+
+        # Assuming isn and Kp are not available in the new format
+        # Calculate isn based on the formula
+        logging.info('Calculating ISN from 10.7cm flux proxy, (isn=int((1.14)*flux_10cm-73.21))')
+        isn = int((1.14) * flux_10cm - 73.21)
+        logging.info('10.7cm flux: %s, proxied ISN: %s', flux_10cm, isn)
+
+    except json.JSONDecodeError as e:
+        logging.error('JSON decode error: %s', e)
+    except ValueError as e:
+        logging.error('Value Error: %s', e)
+    except Exception as e:
+        logging.error('Extract data, other error: %s', e)
+    else:
+        return isn, flux_10cm, epoch, esdate_object
 
 # Build and post payload to Splunk HEC
 
@@ -263,7 +256,7 @@ def hec_send(isn, flux_10cm, Kp, date):
     except requests.exceptions.HTTPError as e:
         logging.error('--- Splunk post failed. Error: %s', e)
     except Exception as e:
-        logging.error('Page load, some other error: %s', e)
+        logging.error('Data load, some other error: %s', e)
     else:
         return r
 
@@ -292,7 +285,7 @@ def es_send(isn, flux_10cm, Kp, date):
     except requests.exceptions.HTTPError as e:
         logging.error('--- Elasticsearch post failed. Error: %s', e)
     except Exception as e:
-        logging.error('Page load, some other error: %s', e)
+        logging.error('Data load, some other error: %s', e)
     else:
         return r
 
@@ -312,7 +305,7 @@ def discord_post(isn, flux_10cm, Kp):
         logging.error(
             '--- Siscord post failed, %s load failed. %s Error: %s', url, r, err)
     except Exception as e:
-        logging.error('Page load, some other error: %s', e)
+        logging.error('Data load, some other error: %s', e)
     else:
         return r
 
@@ -331,11 +324,13 @@ async def test(request: Request, status_code=200, username: str = Depends(getAut
 @app.get("/api/v1/isn")
 @app.get("/v1/isn")
 async def home(request: Request):
-    logging.info('Solar data enpoint response to client: %s',
+    logging.info('Solar data response to client: %s',
                  request.client.host)
-    page = get_page(isn_url)
-    isn, flux_10cm, Kp, epoch, date = extract_data(page)
-    daily = dict(isn=isn, flux_10cm=flux_10cm, Kp=Kp, date=date)
+    ISN_DATA = get_data(ISN_URL)
+    KP_DATA = get_data(KP_URL)
+    isn, flux_10cm, epoch, date = extract_isn(ISN_DATA)
+    kp, epoch, date = extract_kp(KP_DATA)
+    daily = dict(isn=isn, flux_10cm=flux_10cm, Kp=kp, date=date)
     response = dict(solar=daily)
     # responseCode = 200
     # return make_response(jsonify(response), responseCode)
@@ -348,3 +343,16 @@ async def health(request: Request):
     logging.info('Health enpoint response to client %s', client)
     hostname = socket.gethostname()
     return dict(status="healthy", hostname=hostname)
+
+# Set up the scheduler
+sched = BackgroundScheduler(daemon=True)
+# Schedule the ingestion every 6 hours
+sched.add_job(func=dailyIsnIngest, trigger='interval', hours=24)
+# Schedule posting to discord every 12 hours
+sched.add_job(func=dailyIsnDiscord, trigger='interval', hours=24)
+# Schedule posting to Blynk
+sched.add_job(func=postBlynk, trigger='interval', minutes=60)
+# start all schedulers
+sched.start()
+# Log all sched jobs
+logging.info(sched.print_jobs(out=sys.stdout))