+++ CHANGELOG.md
@@ -2,6 +2,147 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.23](https://github.com/apider-coding/isn-poller/compare/v1.0.22...v1.0.23) (2025-10-02)
+
+
+### Bug Fixes
+
+* some cleanup and logging ([fc568e8](https://github.com/apider-coding/isn-poller/commit/fc568e8ee17db63c0fcb0ae05713954c5996c1c0))
+
+### [1.0.22](https://github.com/apider-coding/isn-poller/compare/v1.0.21...v1.0.22) (2025-10-02)
+
+
+### Bug Fixes
+
+* use kp datetime ([1756ef7](https://github.com/apider-coding/isn-poller/commit/1756ef7c300344e768cbdfee470ac269ad6048f4))
+
+### [1.0.21](https://github.com/apider-coding/isn-poller/compare/v1.0.20...v1.0.21) (2025-10-02)
+
+
+### Bug Fixes
+
+* cleanup ([d08ba77](https://github.com/apider-coding/isn-poller/commit/d08ba77738bdd0ac4e0efa14faec465266db1863))
+
+### [1.0.20](https://github.com/apider-coding/isn-poller/compare/v1.0.19...v1.0.20) (2025-10-02)
+
+
+### Bug Fixes
+
+* update api ([9aca420](https://github.com/apider-coding/isn-poller/commit/9aca420c4c6d8fab9f6d3c7ae05bc29df77b23ef))
+
+### [1.0.19](https://github.com/apider-coding/isn-poller/compare/v1.0.18...v1.0.19) (2025-10-02)
+
+
+### Bug Fixes
+
+* add Kp data ([dd5557e](https://github.com/apider-coding/isn-poller/commit/dd5557effa90ea743bef2f3ddaa542981efb6927))
+
+### [1.0.18](https://github.com/apider-coding/isn-poller/compare/v1.0.17...v1.0.18) (2025-10-01)
+
+
+### Bug Fixes
+
+* log time tag ([467173f](https://github.com/apider-coding/isn-poller/commit/467173f0ed2b143230d7e96e332571fbea5f5e6f))
+
+### [1.0.17](https://github.com/apider-coding/isn-poller/compare/v1.0.16...v1.0.17) (2025-10-01)
+
+
+### Bug Fixes
+
+* remoce sending to splunk ([d0d5cb3](https://github.com/apider-coding/isn-poller/commit/d0d5cb3521c6c5a894bfdccdabd53c3c9296a5e5))
+
+### [1.0.16](https://github.com/apider-coding/isn-poller/compare/v1.0.15...v1.0.16) (2025-10-01)
+
+
+### Bug Fixes
+
+* logging ([b76e2c4](https://github.com/apider-coding/isn-poller/commit/b76e2c499339061f4001ccaf170ab618f78086ae))
+
+### [1.0.15](https://github.com/apider-coding/isn-poller/compare/v1.0.14...v1.0.15) (2025-10-01)
+
+### [1.0.14](https://github.com/apider-coding/isn-poller/compare/v1.0.13...v1.0.14) (2025-10-01)
+
+### [1.0.13](https://github.com/apider-coding/isn-poller/compare/v1.0.12...v1.0.13) (2025-10-01)
+
+
+### Bug Fixes
+
+* update interval ([ab01fa5](https://github.com/apider-coding/isn-poller/commit/ab01fa5353592f5b0e9d480151d8efa564166e7d))
+
+### [1.0.12](https://github.com/apider-coding/isn-poller/compare/v1.0.11...v1.0.12) (2025-10-01)
+
+
+### Bug Fixes
+
+* typo ([364614a](https://github.com/apider-coding/isn-poller/commit/364614a1493e1dc552df65d75cc9153243134e25))
+
+### [1.0.11](https://github.com/apider-coding/isn-poller/compare/v1.0.10...v1.0.11) (2025-10-01)
+
+
+### Bug Fixes
+
+* add version output ([92821f9](https://github.com/apider-coding/isn-poller/commit/92821f98c4903bcf71f5751c9bde44c39a62a1f9))
+* une min ([af3d065](https://github.com/apider-coding/isn-poller/commit/af3d065baaff0a09fbb500759eb00295e3d2e13e))
+
+### [1.0.10](https://github.com/apider-coding/isn-poller/compare/v1.0.9...v1.0.10) (2025-10-01)
+
+
+### Bug Fixes
+
+* bump python to 3.9 ([d50b12c](https://github.com/apider-coding/isn-poller/commit/d50b12c7c1833f18bc62c72bc5731638762d1259))
+
+### [1.0.9](https://github.com/apider-coding/isn-poller/compare/v1.0.8...v1.0.9) (2025-10-01)
+
+### [1.0.8](https://github.com/apider-coding/isn-poller/compare/v1.0.7...v1.0.8) (2025-10-01)
+
+
+### Bug Fixes
+
+* cleanup ([7f54cc8](https://github.com/apider-coding/isn-poller/commit/7f54cc82e8a6d384c5b29a0b584b815e39de2da8))
+
+### [1.0.7](https://github.com/apider-coding/isn-poller/compare/v1.0.6...v1.0.7) (2025-10-01)
+
+
+### Bug Fixes
+
+* Using the 10.7cm Flux as a Proxy for Sunspot Number ([16e3c5a](https://github.com/apider-coding/isn-poller/commit/16e3c5a17d84ef22e1a19507827f8868d14f9255))
+
+### [1.0.6](https://github.com/apider-coding/isn-poller/compare/v1.0.5...v1.0.6) (2025-10-01)
+
+
+### Bug Fixes
+
+* Convert metric values to strings in Blynk URL construction ([a61c8cc](https://github.com/apider-coding/isn-poller/commit/a61c8cc14f515e438566f09ddf465e859ea19aa4))
+
+### [1.0.5](https://github.com/apider-coding/isn-poller/compare/v1.0.4...v1.0.5) (2025-10-01)
+
+
+### Bug Fixes
+
+* flux to integer ([fc82c4e](https://github.com/apider-coding/isn-poller/commit/fc82c4e7b4aba75f31c9323602810bd9e4f01843))
+* turn off dependabot ([a7069b8](https://github.com/apider-coding/isn-poller/commit/a7069b83c7925bf4a5562055e212a474aa97c37a))
+
+### [1.0.4](https://github.com/apider-coding/isn-poller/compare/v1.0.3...v1.0.4) (2025-10-01)
+
+
+### Bug Fixes
+
+* rename var ([639f593](https://github.com/apider-coding/isn-poller/commit/639f5937b92d5de0c231a5634d928ee7c8e19ff6))
+
+### [1.0.3](https://github.com/apider-coding/isn-poller/compare/v1.0.2...v1.0.3) (2025-10-01)
+
+
+### Bug Fixes
+
+* remove json ([f7a8e0b](https://github.com/apider-coding/isn-poller/commit/f7a8e0b5a5c32afd3cd248d84dc49115c5a04e72))
+
+### [1.0.2](https://github.com/apider-coding/isn-poller/compare/v1.0.1...v1.0.2) (2025-10-01)
+
+
+### Bug Fixes
+
+* update data getting ([4f00049](https://github.com/apider-coding/isn-poller/commit/4f00049b9416ca8b0deea83a8c22fade4b3b1f49))
+
 ### [1.0.1](https://github.com/apider-coding/isn-poller/compare/v1.0.0...v1.0.1) (2023-05-30)